Code generation moved to its own project, moves zigbee types/commands to subcrate
The code generation part became non trivial. Too much too review and
have the ziggurat project be responsible for. Furthermore the macro
became general enough that it might benefit other projects. Therefore I
have spun it off into its own crate. See:
https://github.com/dvdsk/abstract-bits
