+++ crates/stack/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod ieee_802154;
 pub mod spinel;
 pub mod spinel_client;
-pub mod types;
 pub mod zigbee_aps;
 pub mod zigbee_nwk;
-pub mod zigbee_nwk_commands;
 pub mod zigbee_stack;
+
+pub use zigbee_parts;
