--- crates/wire-format-derive/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "wire-format-derive"
-version = "0.1.0"
-rust-version.workspace = true
-edition.workspace = true
-license.workspace = true
-authors.workspace = true
-
-[dependencies]
-syn = { version = "2", features = ["full", "extra-traits"] }
-quote = "1"
-proc-macro2 = "1" 
-proc-macro-error2 = "2"
-
-[lib]
-proc-macro = true
