--- crates/wire-format-derive/src/errors.rs
@@ -1,1 +1,0 @@
-
