+++ crates/zigbee-parts/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "zigbee-parts"
+version = "0.1.0"
+rust-version.workspace = true
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+
+[dependencies]
+# will be replaced with crates.io dependency later
+abstract-bits = { git = "https://github.com/dvdsk/abstract-bits" }
+num_enum = "0.7.3"
+hex = "0.4.3"
+thiserror = "2.0.12"
