+++ crates/zigbee-parts/src/legacy_members.rs
